moved fetch to usersList
