+++ client/src/views/AdminDashboardView/AdminDashboardView.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import UsersList from './UsersList/UsersList';
 import Tabs from '../../components/Tabs/Tabs';
-import User from './User/User';
 import './AdminDashboardView.css';
 
 class AdminDashboardView extends React.Component {
@@ -10,31 +9,13 @@ class AdminDashboardView extends React.Component {
 
         this.state = {
             tabTitle: this.tabTitles[0],
+            usersList: [],
         };
     }
 
     tabTitles = ['Users', 'Forms'];
-    users = [];
-    tabComponents = [<UsersList usersList={this.users}/>, <p>Forms</p>];
 
-    componentDidMount() {
-        fetch ('/get_users', {
-            method: 'GET',
-            credentials: 'include',
-        }).then((response) => {
-            if (response.ok) {
-                return response.json()
-            } else {
-                throw new Error('Invalid credentials');
-            }
-        }).then((data) => {
-            data.forEach(user => {
-                this.users.push(new User(user.name, user.email, 0));
-            });
-        }).catch((err) => {
-            console.error(err);
-        })
-    }
+    
 
     handleTabChange = (index, title) => {
         this.setState({tabTitle: title})
@@ -50,7 +31,7 @@ class AdminDashboardView extends React.Component {
                 </div>
                 <Tabs
                     titles={this.tabTitles}
-                    components={this.tabComponents}
+                    components={[<UsersList />, <p>Forms</p>]}
                     onTabChangeCallback={this.handleTabChange}
                 />
             </div>
